Using same client functions approach, moved pools to flat structure
