+++ client/pool.go
@@ -1,7 +1,7 @@
 package client
 
 import (
-	"context"
+	"encoding/json"
 	"fmt"
 	"net/http"
 
@@ -9,17 +9,27 @@ import (
 )
 
 func (c *Client) GetPool(name string) (*models.PoolJson, error) {
-	var pool models.PoolJson
-	// Do GET request
-	endpoint := fmt.Sprintf("/agentPools/name:%s", name)
-	resp, err := c.GetRequest(context.Background(), endpoint, "", &pool)
 
+	endpoint := fmt.Sprintf("%s/agentPools/name:%s", c.RestURL, name)
+
+	req, err := http.NewRequest("GET", endpoint, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	resp, err := c.request(req)
 	if err != nil {
 		return nil, err
 	}
 	if resp.StatusCode == http.StatusNotFound {
 		return nil, nil
 	}
-	// Return pool
-	return &pool, nil
+
+	actual := models.PoolJson{}
+	err = json.Unmarshal(resp.Body, &actual)
+	if err != nil {
+		return nil, err
+	}
+
+	return &actual, nil
 }
