+++ teamcity/provider.go
@@ -11,7 +11,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"os"
 	"terraform-provider-teamcity/client"
-	"terraform-provider-teamcity/teamcity/pool"
 )
 
 var (
@@ -154,7 +153,7 @@ func (p *teamcityProvider) DataSources(_ context.Context) []func() datasource.Da
 	return []func() datasource.DataSource{
 		NewServerDataSource,
 		NewBuildConfDataSource,
-		pool.NewPoolDataSource,
+		NewPoolDataSource,
 	}
 }
